+++ mrmustard/lab/abstract/measurement.py
@@ -20,10 +20,10 @@ from abc import ABC, abstractmethod
 from typing import Iterable, Sequence, Union
 
 from mrmustard import settings
+from mrmustard.math.parameter_set import ParameterSet
+from mrmustard.math.parameters import Constant, Variable
 from mrmustard.utils.typing import Tensor
-
 from .state import State
-
 import mrmustard.math as math
 
 
@@ -46,6 +46,25 @@ class Measurement(ABC):
         self._is_postselected = False if outcome is None else True
         """used to evaluate if the measurement outcome should be
         sampled or is already defined by the user (postselection)"""
+
+        self._parameter_set = ParameterSet()
+
+    def _add_parameter(self, parameter: Union[Constant, Variable]):
+        r"""
+        Adds a parameter to a transformation.
+
+        Args:
+            parameter: The parameter to add.
+        """
+        self.parameter_set.add_parameter(parameter)
+        self.__dict__[parameter.name] = parameter
+
+    @property
+    def parameter_set(self):
+        r"""
+        The set of parameters for this transformation.
+        """
+        return self._parameter_set
 
     @property
     def modes(self):