+++ mrmustard/physics/gaussian.py
@@ -164,8 +164,8 @@ def squeezing_symplectic(r: Union[Scalar, Vector], phi: Union[Scalar, Vector]) -
     Returns:
         Tensor: symplectic matrix of a squeezing gate
     """
-    r = math.atleast_1d(r)
-    phi = math.atleast_1d(phi)
+    r = math.atleast_1d(r, math.float64)
+    phi = math.atleast_1d(phi, math.float64)
     if r.shape[-1] == 1:
         r = math.tile(r, phi.shape)
     if phi.shape[-1] == 1:
@@ -195,8 +195,8 @@ def displacement(x: Union[Scalar, Vector], y: Union[Scalar, Vector]) -> Vector:
     Returns:
         Vector: displacement vector of a displacement gate
     """
-    x = math.atleast_1d(x)
-    y = math.atleast_1d(y)
+    x = math.atleast_1d(x, math.float64)
+    y = math.atleast_1d(y, math.float64)
     if x.shape[-1] == 1:
         x = math.tile(x, y.shape)
     if y.shape[-1] == 1:
@@ -288,11 +288,11 @@ def two_mode_squeezing_symplectic(r: Scalar, phi: Scalar) -> Matrix:
     Returns:
         Matrix: symplectic matrix of a two-mode squeezing gate
     """
-    cp = math.cos(phi)
-    sp = math.sin(phi)
-    ch = math.cosh(r)
-    sh = math.sinh(r)
-    zero = math.zeros_like(r)
+    cp = math.cast(math.cos(phi), math.float64)
+    sp = math.cast(math.sin(phi), math.float64)
+    ch = math.cast(math.cosh(r), math.float64)
+    sh = math.cast(math.sinh(r), math.float64)
+    zero = math.cast(math.zeros_like(r), math.float64)
     return math.astensor(
         [
             [ch, cp * sh, zero, sp * sh],
@@ -809,7 +809,7 @@ def fidelity(mu1: Vector, cov1: Matrix, mu2: Vector, cov2: Matrix) -> float:
 
     _fidelity = f0 * math.exp((-1 / 2) * dot)  # square of equation 95
 
-    return math.cast(_fidelity, "float64")
+    return math.real(_fidelity)
 
 
 def physical_partial_transpose(cov: Matrix, modes: Sequence[int]) -> Matrix: